task: 0.5, Step 5: If there are still unresolved issues, use ask_human to request assistance from a human.

+++ src/main/java/org/eclipse/cargotracker/interfaces/Coordinates.java
@@ -19,7 +19,7 @@ public class Coordinates {
   }
 
   public Float getLatitudeFloat() {
-    return Float.valueOf(latitude);
+    return Float.valueOf((float)latitude);
   }
   
   public double getLongitude() {
